+++ CaptivityEvents.csproj
@@ -116,8 +116,8 @@
     <Compile Include="Notifications\CEEventMapNotificationItemVM.cs" />
     <Compile Include="Helper\CEHelper.cs" />
     <Compile Include="Notifications\CESceneNotification.cs" />
-    <Compile Include="Patches\CEBKAttributes.cs" />
-    <Compile Include="Patches\CEHeroCreatorPatch.cs" />
+    <Compile Include="Patches\CEBKPatch.cs" />
+    <Compile Include="Patches\CEPatchHeroCreator.cs" />
     <Compile Include="Patches\CEPatchAttributes.cs" />
     <Compile Include="Patches\CEPatchBasicCharacterObject.cs" />
     <Compile Include="Patches\CEPatchCharacterObject.cs" />
